harden idempotent no-op flows

+++ internal/modules/project/usecase/project_usecase.go
@@ -2,11 +2,13 @@ package usecase
 
 import (
 	"context"
+	"errors"
 
 	"github.com/Roisfaozi/go-clean-boilerplate/internal/modules/project/entity"
 	"github.com/Roisfaozi/go-clean-boilerplate/internal/modules/project/model"
 	"github.com/Roisfaozi/go-clean-boilerplate/internal/modules/project/repository"
 	"github.com/Roisfaozi/go-clean-boilerplate/pkg/exception"
+	"gorm.io/gorm"
 )
 
 type ProjectUseCase interface {
@@ -86,6 +88,16 @@ func (u *projectUseCase) UpdateProject(ctx context.Context, id string, req model
 }
 
 func (u *projectUseCase) DeleteProject(ctx context.Context, id string) error {
+	project, err := u.repo.GetByID(ctx, id)
+	if err != nil {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return nil
+		}
+		return err
+	}
+	if project == nil {
+		return nil
+	}
 	return u.repo.Delete(ctx, id)
 }
